Add timestamp check to usage of local config when RUNLOCAL is set

+++ jenkins.sh
@@ -4,25 +4,27 @@
 
 if [ -v RUNLOCAL ]; then
     echo "Running from local mount..."
-    if [ -f "$JENKINS_PERSIST/local/config.xml" ]; then
-        if [ -f "/config.xml" ]; then
-            mv /config.xml /config.xml.bak
+    if [ "$JENKINS_PERSIST/local/config.xml" -nt /config.xml ]; then
+        if [ -f "$JENKINS_PERSIST/local/config.xml" ]; then
+            if [ -f "/config.xml" ]; then
+                mv /config.xml /config.xml.bak
+            fi
+
+            if [ -f "$JENKINS_PERSIST/config.xml" ]; then
+                rm $JENKINS_PERSIST/config.xml
+            fi
+
+            cp $JENKINS_PERSIST/local/config.xml /config.xml
         fi
 
-        if [ -f "$JENKINS_PERSIST/config.xml" ]; then
-            rm $JENKINS_PERSIST/config.xml
+        if [ -f "$JENKINS_PERSIST/local/jenkins.plugins" ]; then
+            if [ -d "/plugins" ]; then
+                rm -rf /plugins
+                mkdir /plugins
+            fi
+
+            /install_plugins.sh /plugins $JENKINS_PERSIST/local/jenkins.plugins
         fi
-
-        cp $JENKINS_PERSIST/local/config.xml /config.xml
-    fi
-
-    if [ -f "$JENKINS_PERSIST/local/jenkins.plugins" ]; then
-        if [ -d "/plugins" ]; then
-            rm -rf /plugins
-            mkdir /plugins
-        fi
-
-        /install_plugins.sh /plugins $JENKINS_PERSIST/local/jenkins.plugins
     fi
 fi
